==> prioritization/pre-processing/Code/0. ProcessProfileLog.ipynb <==
import os


data_dir = r"H:\Data\PLOVER_TEST_LOG\HBaseTestLogs"
output_file = r"H:\Data\PLOVER_TEST_LOG\hbase_test_log_summary0.log"

output = open(output_file, "w")
for root, dirs, files in os.walk(data_dir, topdown=False):
    for name in files:
        real_path = os.path.join(root, name).replace("\\", "/")
        print(real_path)
        with open(real_path, "r", encoding="utf-8", errors="ignore") as log_file:
            for line in log_file.readlines():
                if "[LoggingOverhead]" in line:
                    time = line[:line.find("[")-1]
                    logging_id = line[line.find("LoggingID:"): line.find(", Time consumption")].replace("LoggingID: ", "")
                    time_cost = line[line.find("Time consumption"):line.find(", Space consumption: ")]\
                        .replace("Time consumption: ", "").replace(" ns", "")
                    space_cost = line[line.find("Space"):].replace("Space consumption: ", "")
                    space_cost = space_cost[:space_cost.find(" bytes")]
#                     output.write("\t".join([name, time, logging_id, time_cost, space_cost]))
                    output.write("\t".join([logging_id, time_cost, space_cost])+"\n")
output.close()

import os

data_dir = r"G:\Data\PLOVER_TEST_LOG\hadoop_test_logs_1\hadoop_test_logs"
for root, dirs, files in os.walk(data_dir, topdown=False):
    for name in files:
        real_path = os.path.join(root, name).replace("\\", "/")
        with open(real_path, "r", encoding="utf-8", errors="ignore") as log_file:
            for line in log_file.readlines():
                if "[LoggingOverhead]" in line:
                    print(real_path)
                    break

import re
with open(r"H:\Data\PLOVER_TEST_LOG\cassandra_test_log_summary0.log", "r") as log_file:
    for line in log_file.readlines():
        logging_ids = re.findall("\w{8}-\w{4}", line)
        if len(logging_ids) == 0:
            print(line)
