==> prioritization/pre-processing/Code/1.1 CalcProfileStat.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

project = "HBase"
index = 0

log_path = r"H:\Data\PLOVER_TEST_LOG\\" + project.lower() + "_test_log_summary" + str(index) + ".log"

data = pd.read_csv(log_path, sep="\t", header=None, names=["LoggingID", "Time(ns)", "Space(bytes)"])

log_ids = np.unique(data["LoggingID"])

len(log_ids)

log_ids

with open("../Data//" + project+"SelectedID.txt", "r") as full_log_id_file:
    full_log_ids = [log_id.strip() for log_id in full_log_id_file.readlines()]

len(full_log_ids)

np.setdiff1d(full_log_ids, log_ids)

np.setdiff1d(log_ids, full_log_ids)

data = data.set_index(["LoggingID"])

data = data.groupby(["LoggingID"])

basic_stat = data.agg(
    {
        'Time(ns)': ['median', 'mean', 'max', 'sum', 'count'],
        'Space(bytes)': ['median', 'mean', 'max', 'sum', 'count']
    }
)

basic_stat.to_csv(project+"NewProfileStat"+str(index)+".csv")
