==> prioritization/pre-processing/Code/1.2 AggrProfileStat.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

data_files = [
    "CassandraNewProfileStat1.csv", "CassandraNewProfileStat2.csv", "CassandraNewProfileStat3.csv",
    "CassandraNewProfileStat4.csv", "CassandraNewProfileStat5.csv", "CassandraNewProfileStat6.csv",
    "CassandraNewProfileStat7.csv", "CassandraNewProfileStat8.csv"
]

# Create a Pandas Excel writer using XlsxWriter as the engine.
writer = pd.ExcelWriter('CassandraNewProfileStatSummary.xlsx', engine='xlsxwriter')

profiles = []
for i, data_file in enumerate(data_files):
    data = pd.read_csv(data_file, header=None, skiprows=[0,1,2], 
                       names=["LoggingID", "Time(ns)Median", "Time(ns)Mean", "Time(ns)Max", "Time(ns)Sum", "Time(ns)Count", 
                              "Space(bytes)Median", "Space(bytes)Mean", "Space(bytes)Max", "Space(bytes)Sum", "Space(bytes)Count"])
    # Convert the dataframe to an XlsxWriter Excel object.
    data.to_excel(writer, sheet_name=str(i), index=False)
    data = data.set_index(["LoggingID"])
    profiles.append(data)
summary = pd.concat(profiles)
summary = summary.groupby(["LoggingID"])
aggr_stat = summary.agg(
    {
        'Time(ns)Count': ['mean'],
        'Time(ns)Sum': ['mean'],
        'Time(ns)Max': ['max']
    }
)
aggr_stat.to_excel(writer, sheet_name="summary")

# Close the Pandas Excel writer and output the Excel file.
writer.save()
