==> prioritization/pre-processing/Code/2.1 CalcMethodExecutionFrequency.ipynb <==
import numpy as np
import pandas as pd
import re
import json

project = "Cassandra"
guards_result_path = r"E:\IDEAWorkspace\plover\docs\result\\" + project + "_Data_Control_SideEffect_Raw_Result.txt"

string = "2021-01-19 13:57:26 [INFO] - Find overhead at method <org.apache.hadoop.hdfs.server.datanode.ShortCircuitRegistry: org.apache.hadoop.hdfs.server.datanode.ShortCircuitRegistry$NewShmInfo createNewMemorySegment(java.lang.String,org.apache.hadoop.net.unix.DomainSocket)>:305 has 16 units and 4 lines"

header_template = re.compile(r'.*Find overhead at method <(\S+).*>:(\w+).*')

header_template.findall(string)

logging_records = []
with open(guards_result_path, "r") as guards_result_file:
    lines = guards_result_file.readlines()
    ids = set()
    for line in lines:
        if "Find overhead at method" in line:
            if len(ids) != 0:
                for pre_id in ids:
                    logging_records.append([file_name, line_num, pre_id])
                    print(file_name, line_num, pre_id)
            ids = set()
            matched = header_template.findall(line)
            class_name = matched[0][0][:-1]
            if "$" in class_name:
                file_name = class_name[class_name.rfind(".")+1:class_name.find("$")]+".java"
            else:
                file_name = class_name[class_name.rfind(".")+1:]+".java"
            line_num = matched[0][1]
        else:
            new_ids = re.findall("\w{8}-\w{4}", line)
            ids.update(new_ids)

method_hit_path = "E:\IDEAWorkspace\\plover\docs\coverage\\" + project.lower() + "\MethodsHit.txt"

with open(method_hit_path, "r") as method_hit_file:
    lines = method_hit_file.readlines()
    raw_hits = []
    for line in lines:
        if ".java" in line:
            raw_hits.append(line.strip().split("\t"))

hits = pd.DataFrame(raw_hits)

hits.columns = ["FileName", "ClassName", "MethodName", "HitCount", "LineNum"]

hits = hits.set_index(["FileName", "LineNum"])

hits

hits.groupby(["FileName", "LineNum"]).agg(['count']).to_csv("test.csv")

#hits.loc["SaslQuorumServerCallbackHandler.java", "52"]["HitCount"]

for record in logging_records:
    exception = True
    
    line_num = int(record[1])
    print("Try to find " + record[0] + " " + record[1])
    while(exception):
        if (line_num < 0):
            # try to increase
            inner_exception = True
            line_num = int(record[1])+1
            while(inner_exception):
                if (line_num > 10000):
                    print("Not found")
                    inner_exception = False
                try:
                    hit_count = hits.loc[record[0], str(line_num)]["HitCount"]
                    print("Found at " + str(line_num))
                    print("MAP", record[2], hit_count[0], sep="\t")
                    inner_exception = False
                except:
                    line_num += 1
            
            exception = False
        try:
            hit_count = hits.loc[record[0], str(line_num)]["HitCount"]
            print("Found at " + str(line_num))
            print("MAP", record[2], hit_count[0], sep="\t")
            exception = False
        except:
            line_num -= 1
