==> prioritization/pre-processing/Code/2.1 CountOfMethodCalls.ipynb <==
import re

project = "Cassandra"
data_dir = r'E:\IDEAWorkspace\plover\docs\result\\' + project + '_Data_Control_SideEffect_Raw_Result.txt'
output_dir = project + '_Count_Of_Method_Calls.txt'

string = "-> if $stack54 == 0 goto (branch) AT LINE 544 with ID"
re.findall("\w{8}-\w{4}", string)

data = {}
with open(data_dir, "r") as data_file:
    for line in data_file.readlines():
        ids = re.findall("\w{8}-\w{4}", line)
        if len(ids) != 0:
            if ("interfaceinvoke" in line or "virtualinvoke" in line or "staticinvoke" in line or "specialinvoke" in line) and \
            ("goto interfaceinvoke" not in line and "goto virtualinvoke" not in line and "goto staticinvoke" not in line and "goto staticinvoke" not in line) and \
            ("isDebugEnabled" not in line and "isTraceEnabled" not in line):
                for log_id in ids:
                    if log_id in data:
                        data[log_id] = data[log_id] + 1
                    else:
                        data[log_id] = 1
            

len(data)

with open(output_dir, "w") as output_file:
    output_file.write("LoggingID\tMethodCalls\n")
    for log_id in data:
        output_file.write(str(log_id) + "\t" + str(data[log_id]) + "\n")
