==> prioritization/pre-processing/Code/2.1 ProcessMetricsDataInLog.ipynb <==
import numpy as np
import pandas as pd

project = "Cassandra"
#data_file = project+"ProfileStat.xlsx"
metrics_file = r"E:\IDEAWorkspace\plover\docs\profiler\\" + project.lower() + "_metrics.log"
output_dir = project + '_Metrics_Data.txt'

#data = pd.read_excel(data_file, sheet_name='ProfileStat')

#data["LoggingID"]

metrics = {}
with open(metrics_file, "r") as metrics_data:
    lines = metrics_data.readlines()
    for line in lines:
        line = line[line.find("LoggingID: "):]
        line = line.replace(": Metrics{potentialMethodCall=", " ").replace(", potentialInstruction=", " ").replace("LoggingID: ", "").replace("}", "").strip()
        entry = line.split(" ")
        metrics[entry[0]] = [entry[1], entry[2]]

with open(output_dir, "w") as output_file:
    output_file.write("LoggingID\tPotentialMethodCalls\tPotentialInstructions\n")
    for logging_id in metrics:
        entry = metrics[logging_id]
        output_file.write(logging_id + "\t" + entry[0] + "\t" + entry[1] + "\n")
