==> prioritization/pre-processing/Code/3. PrepareLtRData.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

# ranker = "time_only"
ranker = "time_and_space_avg"

project = "Cassandra"
output_dir = r"E:\IDEAWorkspace\plover\prioritization\ltr\\" + project + r"Data\folds5"
profile_data_file = project+"NewProfileStatSummary.xlsx"
metrics_data_file = project+"_Metrics_Data.txt"
instructions_data_file = project+"_Count_Of_Instructions.txt"
methodcalls_data_file = project+"_Count_Of_Method_Calls.txt"
methodfreq_data_file = project+"_Method_Frequency.txt"

profile_data = pd.read_excel(profile_data_file, sheet_name="summary", index_col=0)
metrics_data = pd.read_csv(metrics_data_file, sep="\t", index_col=0)
instructions_data = pd.read_csv(instructions_data_file, sep="\t", index_col=0)
methodcalls_data = pd.read_csv(methodcalls_data_file, sep="\t", index_col=0)
methodfreq_data = pd.read_csv(methodfreq_data_file, sep="\t", index_col=0)

# methodcalls_data

# join data by LoggingID in profile_data
joined_data = pd.concat([profile_data, metrics_data, instructions_data, methodcalls_data, methodfreq_data], axis=1, join="inner")

joined_data.columns

if ranker == "time_only":
    # sort by Time(ns)Sum (mean of multiple runs of Time(ns)Sum)
    sorted_data = joined_data.sort_values(by=['Time(ns)Sum'])
    # Set ReverseRank, will be used as the relevance score in LearnToRank
    sorted_data["ReverseRank"] = np.arange(1, len(sorted_data)+1)
else:
    sorted_data = joined_data.sort_values(by=['Time(ns)Sum'])
    sorted_data["TimeReverseRank"] = np.arange(1, len(sorted_data)+1)
    sorted_data = sorted_data.sort_values(by=['Space(bytes)Sum'])
    sorted_data["SpaceReverseRank"] = np.arange(1, len(sorted_data)+1)
    sorted_data["ReverseRankTmp"] = sorted_data["TimeReverseRank"] + sorted_data["SpaceReverseRank"]
    sorted_data = sorted_data.sort_values(by=['ReverseRankTmp', "TimeReverseRank"])
    sorted_data["ReverseRank"] = np.arange(1, len(sorted_data)+1)

# Shuffle data
shuffled_data = sorted_data.sample(frac=1)

shuffled_data

shuffled_data.to_csv(project+"FullData.csv")

splited_data = np.array_split(shuffled_data, 5)

for i, split in enumerate(splited_data):
    with open(os.path.join(output_dir, "fold"+str(i)), "w") as fold_file:
        for index, row in split.iterrows():
            fold_file.write("{rel} pid:1 1:{PotentialMethodCalls} 2:{PotentialInstructions} 3:{Instructions} 4:{MethodCalls} 5:{Count} 6:{MethodFreq} # {LoggingID}\n".format(
                rel=row["ReverseRank"], PotentialMethodCalls=row["PotentialMethodCalls"], PotentialInstructions=row["PotentialInstructions"], 
                Instructions=row["Instructions"], MethodCalls=row["MethodCalls"], Count=row["Time(ns)Count"], MethodFreq=row["MethodFrequency"], LoggingID=index
            ))
